move layout to new file
